rs5002mqtt/requirements.txt
hidapi
paho-mqtt==1.6.1
